Release 0.1.2: upgrade git2 for security
This upgrades the [git2] dependency to 0.17.2 to fix a [security vulnerability
in its handling of SSH keys][GHSA-m4ch-rfv5-x5g3]. This was unlikely to affect
iai-parse since it doesn’t fetch data from, or otherwise interact with, remote
repositories.

This vulnerability was discovered months ago, but somehow I missed that
iai-parse could be affected. I actually updated [git-status-vars] in
January to fix it (though it was likely not vulnerable either).

[git2]: https://crates.io/crates/git2
[GHSA-m4ch-rfv5-x5g3]: https://github.com/rust-lang/git2-rs/security/advisories/GHSA-m4ch-rfv5-x5g3
[git-status-vars]: https://github.com/danielparks/git-status-vars