Use cargo-lints to centrally configure lints
Previously lints had to be configured at the top of every entry point in
rust code. This allows lints to be centrally configured in `lints.toml`
and ensures they are applied to all targets.

This also updates the lint list to match my other projects. It includes
a bunch more restriction lints.

To run lints:

    cargo lints clippy --all-targets --all-features

This requires [cargo-lints] to be installed.

[cargo-lints]: https://crates.io/crates/cargo-lints